koreainvestment/requirements.txt
requests>=2.32.0
websockets>=12.0
PyYAML>=6.0.1
pycryptodome>=3.20.0
pandas>=2.2.0
